Moved MavenCoordinate to base-analyzer
