+++ analyzer/base-analyzer/src/test/java/eu/fasten/analyzer/baseanalyzer/MavenCoordinateTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package eu.fasten.analyzer.javacgwala.data;
+package eu.fasten.analyzer.baseanalyzer;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -28,7 +28,6 @@ import java.nio.charset.StandardCharsets;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-
 
 public class MavenCoordinateTest {
