+++ analyzer/base-analyzer/src/test/resources/testpom.txt
