+++ analyzer/javacg-opal/src/test/java/eu/fasten/analyzer/javacgopal/OPALPluginTest.java
@@ -22,7 +22,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import eu.fasten.analyzer.javacgopal.data.MavenCoordinate;
+import eu.fasten.analyzer.baseanalyzer.MavenCoordinate;
 import eu.fasten.analyzer.javacgopal.data.callgraph.PartialCallGraph;
 import java.io.FileNotFoundException;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
@@ -90,7 +90,7 @@ public class OPALPluginTest {
     }
 
     @Test
-    public void testEmptyCallGraph() {
+    public void testEmptyCallGraph() throws JSONException {
         JSONObject emptyCGCoordinate = new JSONObject("{\n"
             + "    \"groupId\": \"activemq\",\n"
             + "    \"artifactId\": \"activemq\",\n"
@@ -106,7 +106,7 @@ public class OPALPluginTest {
     }
 
     @Test
-    public void testFileNotFoundException() {
+    public void testFileNotFoundException() throws JSONException {
         JSONObject noJARFile = new JSONObject("{\n"
             + "    \"groupId\": \"com.visionarts\",\n"
             + "    \"artifactId\": \"power-jambda-pom\",\n"
@@ -123,7 +123,7 @@ public class OPALPluginTest {
     }
 
     @Test
-    public void testNullPointerException() {
+    public void testNullPointerException() throws JSONException {
         JSONObject mvnCoordinate = new JSONObject("{\n"
             + "    \"groupId\": \"ch.epfl.scala\",\n"
             + "    \"artifactId\": \"collection-strawman_0.6\",\n"
