+++ analyzer/javacg-wala/pom.xml
@@ -32,7 +32,13 @@
             <groupId>eu.fasten</groupId>
             <artifactId>core</artifactId>
             <version>0.0.1-SNAPSHOT</version>
-            <!--            <scope>provided</scope>-->
+            <!--<scope>provided</scope>-->
+        </dependency>
+        <dependency>
+            <groupId>eu.fasten.analyzer</groupId>
+            <artifactId>base-analyzer</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <!--<scope>provided</scope>-->
         </dependency>
         <dependency>
             <groupId>com.ibm.wala</groupId>
@@ -59,12 +65,6 @@
             <artifactId>gradle-tooling-api</artifactId>
             <version>4.3</version>
         </dependency>
-        <dependency>
-            <groupId>org.jboss.shrinkwrap.resolver</groupId>
-            <artifactId>shrinkwrap-resolver-impl-maven-embedded</artifactId>
-            <version>3.1.3</version>
-            <type>pom</type>
-        </dependency>
 
         <!-- Uncomment the slf4j package only if you wanna use WALA module outside the FASTEN server -->
 <!--        <dependency>-->
@@ -77,17 +77,6 @@
 <!--            <artifactId>slf4j-api</artifactId>-->
 <!--            <version>1.7.30</version>-->
 <!--        </dependency>-->
-
-        <dependency>
-            <groupId>org.dom4j</groupId>
-            <artifactId>dom4j</artifactId>
-            <version>2.1.1</version>
-        </dependency>
-        <dependency>
-            <groupId>jaxen</groupId>
-            <artifactId>jaxen</artifactId>
-            <version>1.2.0</version>
-        </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
@@ -118,16 +107,6 @@
                             <minimizeJar>false</minimizeJar>
                             <keepDependenciesWithProvidedScope>false
                             </keepDependenciesWithProvidedScope>
-                            <!--                            <createDependencyReducedPom>false</createDependencyReducedPom>-->
-                            <!--                            <artifactSet>-->
-                            <!--&lt;!&ndash;                                <excludes>&ndash;&gt;-->
-                            <!--&lt;!&ndash;                                    <exclude>com.typesafe.play:play-json_2.11:*</exclude>&ndash;&gt;-->
-                            <!--&lt;!&ndash;                                    <exclude>com.typesafe.play:play-functional_2.11:*</exclude>&ndash;&gt;-->
-                            <!--&lt;!&ndash;                                </excludes>&ndash;&gt;-->
-                            <!--                                <includes>-->
-                            <!--                                    <include>*:*</include>-->
-                            <!--                                </includes>-->
-                            <!--                            </artifactSet>-->
                             <transformers>
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
