+++ analyzer/pom.xml
@@ -28,6 +28,7 @@
         <module>javacg-opal</module>
         <module>dummy-analyzer</module>
         <module>metadata-plugin</module>
+        <module>base-analyzer</module>
     </modules>
 
     <build>
